requirements.txt
datasets==3.2.0
google-generativeai==0.8.3
gradio==5.9.0
psycopg[binary]==3.2.3
python-dotenv==1.0.1
PyYAML==6.0.1
sentence-transformers==3.3.1
tqdm==4.67.1
